+++ build.gradle.kts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import eu.cloudnetservice.cloudnet.gradle.applyDefaultJavadocOptions
 import eu.cloudnetservice.cloudnet.gradle.plugins.configureFor
 import eu.cloudnetservice.cloudnet.gradle.plugins.updater.lenientView
 import eu.cloudnetservice.cloudnet.gradle.util.CustomConfigurations
 import eu.cloudnetservice.cloudnet.gradle.util.Versions
+import eu.cloudnetservice.cloudnet.gradle.util.applyDefaultJavadocOptions
 
 plugins {
   id("cloudnet")
